+++ assets/javascripts/kitten/components/form/input/password-input/index.js
@@ -1,40 +1,7 @@
 import React, { useState } from 'react'
-import styled from 'styled-components'
+import PropTypes from 'prop-types'
 import { TextInput } from '../../../form/input/text-input'
 import { PasswordIcon } from '../../../graphics/icons/password-icon'
-import COLORS from '../../../../constants/colors-config'
-import PropTypes from 'prop-types'
-import { pxToRem } from '../../../../helpers/utils/typography'
-
-const StyledPasswordInput = styled.div`
-  position: relative;
-  display: flex;
-
-  .k-Form-PasswordInput__input {
-    padding-right: ${pxToRem(40)};
-  }
-
-  .k-Form-PasswordInput__icon {
-    display: flex;
-    position: absolute;
-    z-index: 1;
-    padding: 0 ${pxToRem(11)};
-    right: 0;
-    top: 0;
-    bottom: 0;
-    cursor: pointer;
-    align-items: center;
-
-    border-radius: var(--border-radius-s);
-
-    &[aria-pressed='true'] {
-      svg {
-        fill: ${COLORS.primary1};
-        transition: all 0.2s;
-      }
-    }
-  }
-`
 
 export const PasswordInput = ({
   name,
@@ -49,23 +16,24 @@ export const PasswordInput = ({
   }
 
   return (
-    <StyledPasswordInput className="k-Form-PasswordInput">
-      <TextInput
-        {...others}
-        name={name}
-        type={isHidden ? 'password' : 'text'}
-        className="k-Form-PasswordInput__input"
-      />
-      <button
-        type="button"
-        aria-label={isHidden ? iconLabel : hiddenIconLabel}
-        onClick={handleClick}
-        aria-pressed={!isHidden}
-        className="k-Form-PasswordInput__icon k-u-reset-button"
-      >
-        <PasswordIcon />
-      </button>
-    </StyledPasswordInput>
+    <TextInput
+      {...others}
+      name={name}
+      type={isHidden ? 'password' : 'text'}
+      has="button"
+      buttonIsInset={true}
+      buttonProps={{
+        'aria-label': isHidden ? iconLabel : hiddenIconLabel,
+        'aria-pressed': !isHidden,
+        onClick: handleClick,
+        className: 'k-PasswordInput__icon',
+        children: (
+          <PasswordIcon
+            color={`var(--color-${isHidden ? 'grey-900' : 'primary-500'})`}
+          />
+        ),
+      }}
+    />
   )
 }