+++ assets/javascripts/kitten/components/layout/grid/index.js
@@ -73,7 +73,7 @@ export const GridCol = ({ className, ...props }) => {
       const col = props[`col-${mediaQuery}`]
       const offset = props[`offset-${mediaQuery}`]
 
-      return classNames(classNamesByMediaQuery, {
+      return classNames({
         [`k-Grid__col--${col}-${mediaQuery}`]: col,
         [`k-Grid__col--offset-${offset}-${mediaQuery}`]:
           !!offset || offset === 0,