+++ assets/javascripts/kitten/config/test-setup.js
@@ -1,6 +1,0 @@
-const chai = require('chai')
-const sinonChai = require('sinon-chai')
-const chaiEnzyme = require('chai-enzyme')
-
-chai.use(chaiEnzyme())
-chai.use(sinonChai)